requirements.txt
django
psycopg2-binary	
django-widget-tweaks
python-decouple
django-qurl-templatetag
Pillow
django-mptt